+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Bug checking if a field stored data on the legacy data table
 - Email Contact Form attribute access in Twig templates
 - Retrieving the current request's page value for paginating ModelTag results
+- Structure Nav tag behavior now respects hidden nodes
 
 ### Changed